requirements.txt
pyyaml
flask
jinja2
